+++ src/components/Breadcrumb/breadcrumb-sep.tsx
@@ -1,0 +1,28 @@
+import classNames from "classnames";
+import type { FC } from "react";
+
+import { useJss } from "@/components/CopyUiProvider";
+import { useTheme } from "@/components/ThemeProvider";
+
+import styles from "./breadcrumb-sep.module.scss";
+
+type BreadcrumbSepProps = {
+  children?: React.ReactNode;
+};
+
+const BreadcrumbSep: FC<BreadcrumbSepProps> = (props) => {
+  const { children = "/" } = props;
+
+  const theme = useTheme();
+  const jss = useJss();
+
+  const stx = jss.hash({
+    "--breadcrumb-sep-color": theme.colors.gray["500"],
+  });
+
+  return <li className={classNames(styles.breadcrumbSep, stx)}>{children}</li>;
+};
+
+BreadcrumbSep.displayName = "Breadcrumb.Sep";
+
+export { BreadcrumbSep };