+++ src/components/Breadcrumb/index.ts
@@ -1,1 +1,3 @@
-export { Breadcrumb } from "./Breadcrumb";
+// Breadcrumb from copy-ui @ 2025-10-26.
+
+export { Breadcrumb } from "./breadcrumb";