+++ src/components/Breadcrumb/CHANGELOG.md
@@ -1,1 +1,2 @@
 - 2025-10-18: Initial version.
+- 2025-10-26: Support `size` prop and custom separator content.